requirements.txt
flask
python-dotenv
google-generativeai
scikit-learn
numpy
